Updated LabBotControl, refining Move2Global
the fnmicon and ikcon where not working as i liked and it was not computing the inverse kinematics properly it was failing heaps. Reverted back to Ikine, added a 2 stage Ikine process, initial guess is accordng to the x an y of the endeffector, then ikine gets the q values of the position then that is used as the next guess for the ikine to work out position and orientation. not working competely yet but 6/8 movement succesfully competeted